+++ src/scripts/CSS.js
@@ -1,0 +1,6 @@
+export const hideDivs = () =>{
+    const auth = document.querySelector(".auth")
+    const loginDiv = document.querySelector(".loginDiv")
+    auth.style.display = "none"
+    loginDiv.style.display = "none"
+  }
