+++ src/scripts/auth/RegisterForm.js
@@ -60,9 +60,3 @@ export const RegisterForm = () => {
   render()
 }
 
-export const hideDivs = () =>{
-  const auth = document.querySelector(".auth")
-  const loginDiv = document.querySelector(".loginDiv")
-  auth.style.display = "none"
-  loginDiv.style.display = "none"
-}
