+++ src/scripts/events/EventForm.js
@@ -21,7 +21,7 @@ export const EventForm = () => {
     <input type="text" id="createForm__eventCity" placeholder="City">
     <h6>State Code</h6>
     <input type="text" id="createForm__eventState" placeholder="State">
-    <br>
+    <br><br>
     <button id="createForm__addEventButton">Add Event</button>`;
     contentTarget.innerHTML = htmlRep;
 }
