+++ src/styles/main.css
@@ -258,6 +258,8 @@ header {
   padding: 75px;
   background-color: white;
   border-radius: 10px;
+  box-shadow: 0px 8px 16px 0px rgba(0,0,0,0.5);
+
 
 }
 footer{
